pyproject.toml
[tool.poetry]
name = "ninjascratch"
version = "0.1.0"
description = "A ground-up Django-Ninja API with no boilerplate"
authors = ["Ryan Palermo"]
readme = "README.md"

[tool.poetry.dependencies]
django = "^4.2.5"
django-configurations = "^2.4.1"
django-extensions = "^3.2.3"
django-ninja = "^0.22.2"
django-ninja-extra = "^0.19.3"
django-ninja-jwt = "^5.2.5"
djangorestframework = "^3.14.0"
ipython = "^8.15.0"
psycopg = {extras = ["binary"], version = "^3.1.10"}
pydantic = "^1.10.12"
python = "3.10.4"
uuid = "^1.30"
uwsgi = "^2.0.20"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}

[tool.poetry.group.development.dependencies]
django-debug-toolbar = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
